Factor out logic for inferring widget from abbreviation
