+++ panel/widgets/__init__.py
@@ -64,6 +64,7 @@ from .tables import DataFrame, Tabulator  # noqa
 from .terminal import Terminal  # noqa
 from .text_to_speech import TextToSpeech, Utterance, Voice  # noqa
 from .texteditor import TextEditor  # noqa
+from .widget import widget  # noqa
 
 __all__ = (
     "Ace",
